removed duplicate normalize_pred
